Don't require hostid file in mmp cleanup

+++ tests/zfs-tests/tests/functional/mmp/cleanup.ksh
@@ -23,11 +23,13 @@
 
 verify_runnable "global"
 
-log_must mv ${HOSTID_FILE}.bak $HOSTID_FILE
-if is_freebsd; then
-	log_must /etc/rc.d/hostid start
-else
-	log_must zgenhostid "$(<$HOSTID_FILE)"
+if [ -e ${HOSTID_FILE}.bak ]; then
+	log_must mv ${HOSTID_FILE}.bak $HOSTID_FILE
+	if is_freebsd; then
+		log_must /etc/rc.d/hostid start
+	else
+		log_must zgenhostid "$(<$HOSTID_FILE)"
+	fi
 fi
 log_must set_tunable64 MULTIHOST_HISTORY $MMP_HISTORY_OFF
